change client

+++ client/src/views/NannyDetail.vue
@@ -65,7 +65,7 @@
                   v-if="loginAs == 'agency' && authorized"
                 >Edit Info</button>
                 <div>
-                  <EditNannyModal></EditNannyModal>
+                  <EditNannyModal :dataNanny="dataNanny"></EditNannyModal>
                 </div>
                 <button
                   class="buttonA"
@@ -93,6 +93,7 @@ import Vuex from "vuex";
 import NavBar from "../components/Navbar";
 import ChildCard from "../components/ChildCard";
 import EditNannyModal from "../components/EditNannyModal";
+import Swal from "sweetalert2";
 
 export default {
   name: "NannyDetail",
@@ -159,7 +160,18 @@ export default {
       this.$store.dispatch("slideNannyCardToTheRight", Number(this.id) + 1);
     },
     deleteNanny() {
-      this.$store.dispatch("deleteNanny", this.id);
+      Swal.fire({
+        title: "Are you sure?",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes"
+      }).then(result => {
+        if (result.value) {
+          this.$store.dispatch("deleteNanny", this.id);
+        }
+      });
     },
     formatPrice(value) {
       let val = (value / 1).toFixed(2).replace(".", ",");
